Improved JSON templates
JSON templates can now take arrays instead of objects in all cases, to
allow for duplicated keys. For example, in a row or column, the
children do not all need unique IDs.

Also, components with only one slot can have components on the top
level. No need to even put the slot name. We’ll figure it out for you!