Sped up ParseTemplate significantly
It’s the most expensive step in the process right now so that’s kind of
a big deal. The idea is that components that are not getting rendered
on the page can defer some of their initialization steps until later.
This may lead to some performance hiccups while the user is actually on
the page (although in most web apps the load will be spread out such
that it’s unnoticeable). Still, load time is king!